# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MegaChad Agent API",
    "version": "1.0.0",
    "description": "Burn-to-create looksmaxxing engine on MegaETH. Burn $MEGACHAD tokens to generate AI-enhanced portraits, mint NFTs, and earn on-chain reputation. All endpoints run on MegaETH (chain ID 4326, RPC: https://mainnet.megaeth.com/rpc).",
    "contact": {
      "name": "MegaChad",
      "url": "https://megachad.xyz"
    }
  },
  "servers": [
    {
      "url": "https://megachad.xyz",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/agent/chat": {
      "post": {
        "operationId": "chatWithMegachad",
        "summary": "Natural language interface for AI agents",
        "description": "Send a plain English message and get back structured, actionable responses with data and suggested next actions. Supports 17 intents: price, stats, wallet, looksmaxx, swap, gallery, leaderboard, bridge, gasless, referral, about, burn, compare, multi_burn, schedule, cross_chain_looksmaxx, register_agent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string", "description": "Natural language query", "example": "What is the current MEGACHAD price?" },
                  "wallet": { "type": "string", "description": "Wallet address for context (0x...)", "pattern": "^0x[a-fA-F0-9]{40}$" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured response with intent, answer, data, and actions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "intent": { "type": "string", "description": "Detected intent (price, stats, wallet, etc.)" },
                    "answer": { "type": "string", "description": "Human-readable answer" },
                    "data": { "type": "object", "description": "Raw data from the underlying API" },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string" },
                          "endpoint": { "type": "string" },
                          "method": { "type": "string" },
                          "body": { "type": "object" }
                        }
                      },
                      "description": "Suggested follow-up actions"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or empty message" }
        }
      },
      "get": {
        "operationId": "getChatInfo",
        "summary": "Get chat endpoint documentation",
        "description": "Returns supported intents, example messages, and endpoint documentation.",
        "responses": {
          "200": {
            "description": "Endpoint documentation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "endpoint": { "type": "string" },
                    "method": { "type": "string" },
                    "description": { "type": "string" },
                    "supportedIntents": { "type": "array", "items": { "type": "string" } },
                    "examples": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/defi/amm/quote": {
      "get": {
        "operationId": "getAmmQuote",
        "summary": "Quote a swap on the MEGACHAD/MEGAGOONER pair",
        "description": "Constant-product math (0.3% fee). Without `from`+`amount`, returns reserves + spot price. With both, returns expected output, price impact bps, and a link to the swap-tx builder. NOTE: MegaChadLP uses tokenA/tokenB rather than the V2 standard token0/token1 — generic V2 routers will not find this pool.",
        "parameters": [
          { "name": "from", "in": "query", "required": false, "schema": { "type": "string", "enum": ["MC", "MG"] } },
          { "name": "amount", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Pool state + optional quote" } }
      }
    },
...
    "/api/defi/amm/add-liquidity-tx": {
      "get": {
        "operationId": "buildAddLiquidityTx",
        "summary": "Build approve+approve+addLiquidity on the MC/MG pair",
        "parameters": [
          { "name": "amountMC", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "amountMG", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "recipient", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "address", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Liquidity provisioning plan" } }
      }
    },
...
  },
...
}